Will after lunch add select box for type. Will thereby finnish "update" part of the CRUD library_items Read is only partly finnished, because I need to figure out if I want to show author for DVD and other things, I will probably allow this bug (feature).
Delete will not even "alert are you sure", it will just delete
Thereafter i will allow for adding library_items
And finally I will check_in and check_out.
Possibly will allow for update of the checked_in if I get time on wednesday.

But currently will not allow update of checked in books?
(can check in books in history and future, so that should be edited by checking out mistakes and correcting the checkin.)