Will now implement on api service and controller: 1. Checkin 2. checkout 3. Update
This order just so that nobody inserts ReferenceBook that is borrowed.
Update does not even change the value "Borrowed" or "isBorrowable".
Or will it when somebody bypasses frontend and sends an update to turn an book into an reference book?
Nahh, if change to reference book i must check an querry so nobody has borrowed it.
Really that is not less safe than the unauthorized version.... where nobody checks "isCheckedIn".

But should return bad request if borrower is not null when switching to reference_book